Implemented genesis and evaluation for GA, added constraints for teachers

+++ src/main/java/com/guu/constraints/Constraint.java
@@ -4,7 +4,7 @@ import com.guu.utils.Timetable;
 
 public abstract class Constraint {
     Boolean hard;
-    public abstract int checkConstraint(Timetable gt); // returns constraint
+    public abstract double checkConstraint(Timetable gt); // returns constraint
     public Boolean getHard() { return hard; }
     protected void setHard(Boolean hard) { this.hard = hard; }
 }
